Add support for flattening prefixes to DynamoDB fields
This makes enables clients to avoid conflicts if flattened schemas happen
to share a field name. The design maintains backwards compatibility with
existing codebases by requiring that users opt-in to this behavior
explicitly. While this increases the mental overhead there is a design
to enable the auto-prefixing as the default behavior when creating the
mapper. However, that is outside of the scope of this current implementation.